requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7
gitdb2==2.0.3             # via gitpython
gitpython==2.1.8
pyyaml==3.12
smmap2==2.0.3             # via gitdb2
